backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
langgraph>=1.0.7
langchain-google-genai>=2.0.0
google-genai>=1.0.0
tavily-python>=0.5.0
jinja2>=3.1.0
tenacity>=9.0.0
pydantic>=2.7.0
python-dotenv>=1.0.0
